fix: controller to langchain refactor
